feat: 引入 Chroma 向量检索并规范环境配置

+++ backend/schema/chat_schema.py
@@ -113,7 +113,7 @@ class RagPreviewChunk(BaseModel):
     """
     file_name: Optional[str] = None  # 命中文本块所属文件名
     chunk_id: int | None = None  # 文本块编号
-    score: int = 0  # 检索分数
+    score: float = 0.0  # 检索分数；关键词模式为整数，向量模式为相似度小数
     source: Optional[str] = None  # 引用来源标识，例如：员工手册.md#chunk-4
     text: str = ""  # 命中的原文片段
     text_preview: str  # 文本预览内容
